requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==3.0.1
attrs==21.2.0
autobahn==22.7.1
Automat==20.2.0
cffi==1.14.6
channels==3.0.5
channels-redis==3.4.1
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
Django==4.1.1
djangorestframework==3.14.0
flake8==5.0.4
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
mccabe==0.7.0
msgpack==1.0.2
pika==1.2.0
psycopg2==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.20
pyflakes==2.5.0
pyOpenSSL==22.1.0
pytz==2021.1
rabbitmq-client==2.4.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==22.8.0
txaio==21.2.1
typing_extensions==4.4.0
zope.interface==5.4.0
